chore: split filter interface file
