+++ packages/shared/components/filters/Filter.svelte
@@ -1,6 +1,6 @@
 <script lang="typescript">
     import { TogglableButton, FilterModal, FilterItem, Modal } from 'shared/components'
-    import type { Filter } from '@core/wallet/interfaces/filter.interface'
+    import type { Filter } from '@core/wallet/interfaces/filter/filter.interface'
     import { deepCopy } from '@lib/utils'
     import type { Writable } from 'svelte/store'
