+++ packages/shared/lib/core/wallet/interfaces/filter/index.ts
@@ -1,0 +1,2 @@
+export * from './date-filter-input.interface'
+export * from './number-filter-input.interface'
