+++ packages/shared/lib/core/wallet/interfaces/filter/number-filter-input.interface.ts
@@ -1,0 +1,12 @@
+export type NumberFilterInput = SingleNumberFilterInput | RangeNumberFilterInput
+
+export type SingleNumberFilterInput = {
+    type: 'single'
+    amount: string
+}
+
+export type RangeNumberFilterInput = {
+    type: 'range'
+    start: string
+    end: string
+}
