+++ packages/shared/lib/core/wallet/stores/selected-account-activities.store.ts
@@ -7,7 +7,7 @@ import { isValueInUnitRange, unitToValue } from '@lib/utils'
 import { Activity } from '../classes/activity.class'
 import { allAccountActivities } from './all-account-activities.store'
 import { isVisibleActivity } from '../utils/isVisibleActivity'
-import { ActivityFilter } from '../interfaces/filter.interface'
+import { ActivityFilter } from '../interfaces/filter/filter.interface'
 import {
     NumberFilterOption,
     BooleanFilterOption,
